bounding poligons updated

+++ app/document/views.py
@@ -11,11 +11,14 @@ BUCKET_NAME = os.environ.get("S3_BUCKET_NAME")
 
 class DocumentListCreateView(generics.ListCreateAPIView):
     serializer_class = DocumentSerializer
-    permission_classes = [permissions.AllowAny]
-    authentication_classes = []
+    permission_classes = [permissions.IsAuthenticated]
+    authentication_classes = [JWTAuthentication]
 
     def get_queryset(self):
-        return Document.objects.all()
+        return Document.objects.filter(user=self.request.user)
+
+    def perform_create(self, serializer):
+        serializer.save(status=Document.Status.PENDING, user=self.request.user)
 
 
 class DocumentRetrieveUpdateDestroyView(generics.RetrieveUpdateDestroyAPIView):
@@ -32,6 +35,7 @@ class DocumentRetrieveUpdateDestroyView(generics.RetrieveUpdateDestroyAPIView):
         return [permissions.AllowAny()]
 
     def perform_destroy(self, instance):
-        s3_loader = S3FileLoaderService(bucket_name=BUCKET_NAME)
-        s3_loader.delete_file(key=instance.s3_key)
+        if instance.s3_key:
+            s3_loader = S3FileLoaderService(bucket_name=BUCKET_NAME)
+            s3_loader.delete_file(key=instance.s3_key)
         instance.delete()
